Sistema/mcp-smosu/requirements.txt
mcp[cli]>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pytest>=8.0.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0